pyproject.toml
[tool.poetry]
name = "quant-ai"
version = "0.0.1"
description = ""
authors = ["J.Stone"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
chromadb = "^0.3.26"
langchain = "^0.0.222"
python-dotenv = "^1.0.0"
gpt4all = "^1.0.1"
llama-cpp-python = "^0.1.67"
urllib3 = "^2.0.3"
PyMuPDF = "^1.22.5"
unstructured = "^0.7.12"
extract-msg = "^0.41.5"
tabulate = "^0.9.0"
pandoc = "^2.3"
pypandoc = "^1.11"
tqdm = "^4.65.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
